OpaEvalRunConfiguration.kt: move Element extention functions to opaplugin/openapiext/ElementExt.kt
