+++ src/main/kotlin/com/github/vgramer/opaplugin/ide/runconfig/OpaEvalRunConfiguration.kt
@@ -5,9 +5,11 @@
 package com.github.vgramer.opaplugin.ide.runconfig
 
 import com.github.vgramer.opaplugin.ide.runconfig.ui.OpaEvalRunCommandEditor
-import com.google.common.annotations.VisibleForTesting
+import com.github.vgramer.opaplugin.openapiext.readPath
+import com.github.vgramer.opaplugin.openapiext.readString
+import com.github.vgramer.opaplugin.openapiext.writePath
+import com.github.vgramer.opaplugin.openapiext.writeString
 import com.intellij.execution.Executor
-import com.intellij.execution.ExternalizablePath
 import com.intellij.execution.configurations.*
 import com.intellij.execution.runners.ExecutionEnvironment
 import com.intellij.openapi.options.SettingsEditor
@@ -16,7 +18,6 @@ import com.intellij.util.execution.ParametersListUtil
 import com.intellij.util.io.isFile
 import org.jdom.Element
 import java.nio.file.Path
-import java.nio.file.Paths
 
 /**
  * the opa eval configuration
@@ -121,30 +122,4 @@ class OpaEvalRunConfiguration(
         element.writeString("addtionalargs", additionalArgs)
 
     }
-}
-
-@VisibleForTesting
-fun Element.writeString(name: String, value: String?) {
-    val opt = Element("option")
-    opt.setAttribute("name", name)
-    opt.setAttribute("value", value ?: "")
-    addContent(opt)
-}
-
-
-private fun Element.readString(name: String): String? =
-    children
-        .find { it.name == "option" && it.getAttributeValue("name") == name }
-        ?.getAttributeValue("value")
-
-@VisibleForTesting
-fun Element.writePath(name: String, value: Path?) {
-    if (value != null) {
-        val s = ExternalizablePath.urlValue(value.toString())
-        writeString(name, s)
-    }
-}
-
-private fun Element.readPath(name: String): Path? {
-    return readString(name)?.let { Paths.get(ExternalizablePath.localPathValue(it)) }
 }
